refactor route exceptions
