+++ src/platform/packages/shared/kbn-typed-react-router-config/index.ts
@@ -8,7 +8,7 @@
  */
 
 export * from './src/create_router';
-export * from './src/invalid_params_exception';
+export * from './src/errors/invalid_route_params_exception';
 export * from './src/encode_path';
 export type * from './src/types';
 export * from './src/outlet';
