+++ src/platform/packages/shared/kbn-typed-react-router-config/src/errors/not_found_route_exception.ts
@@ -7,17 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { last } from 'lodash';
-import { useMatchRoutes } from './use_match_routes';
-import { useRouter } from './use_router';
-import { NotFoundRouteException } from './errors';
-
-export function useRoutePath() {
-  const lastRouteMatch = last(useMatchRoutes());
-  const router = useRouter();
-  if (!lastRouteMatch) {
-    throw new NotFoundRouteException('No route was matched');
+export class NotFoundRouteException extends Error {
+  constructor(message: string) {
+    super(message);
   }
-
-  return router.getRoutePath(lastRouteMatch.route);
 }
